Changing the User Model to authenticate by email.
