+++ newspaper/models.py
@@ -1,5 +1,5 @@
 from django.db import models
-from django.contrib.auth.models import User
+from django.contrib.auth.models import User, AbstractUser
 from time import timezone
 from django.contrib import admin
 from django.db.models.signals import post_save
@@ -36,3 +36,11 @@ class Article(models.Model):
         return self.body[:100] + '...'
 
     
+class User(AbstractUser):
+    name = models.CharField(max_length=255)
+    email = models.CharField(max_length=255, unique=True)
+    password = models.CharField(max_length=255)
+    username = None
+    
+    USERNAME_FIELD = 'email'
+    REQUIRED_FIELDS = []
