+++ newspaper/urls.py
@@ -1,18 +1,18 @@
-from django.urls import path, include
-from . import views
+from django.urls import path
+from newspaper import views
 
 urlpatterns = [
     path('', views.homeView, name='home-view'),
     
-    path('verify/', views.verifyUserExists, name='verify-user'),
-    path('login/', views.loginUser, name='login-user'),
-    path('register/', views.registerUser, name='register-user'),
+    path('verify/', views.user_views.verifyUserExists, name='verify-user'),
+    path('login/', views.user_views.loginUser, name='login-user'),
+    path('register/', views.user_views.registerUser, name='register-user'),
     
     
-    path('articles/', views.articlesView, name="articles-view"),
-    path('articles/<str:id>/', views.singleArticleView, name='single-article-view'),
+    path('articles/', views.article_views.articlesView, name="articles-view"),
+    path('articles/<str:id>/', views.article_views.singleArticleView, name='single-article-view'),
     
-    path('topics/', views.topicView, name='topic-view'),
+    path('topics/', views.article_views.topicView, name='topic-view'),
     
-    path('creators/<str:username>/', views.creatorView, name='creator-view'),
+    path('creators/<str:username>/', views.article_views.creatorView, name='creator-view'),
 ]
